pgw#1600: `Shape` lives with the demand vocabulary, not beside the extractor
An author writing a lane writes one import line — `from gen_worker.demand
import GiB, MiB, Shape, const, per_mp_batch` — because `Shape(pixels=...)` is
part of the DECLARATION vocabulary, not part of the machinery that reads it.
`demand_envelope` keeps the extraction (field-name vocabulary, envelope
derivation) and re-exports the marker for anyone already importing it there.
